+++ on-fe/app/(auth)/login/additional-info/page.js
@@ -1,5 +1,5 @@
 'use client';
-import AdditionalInfoForm from '@/components/auth/AdditionalInfoForm';
+import AdditionalInfoForm from '@/components/form/AdditionalInfoForm';
 import FormLayout from '@/components/form/FormLayOut';
 import '@/app/(auth)/login/additional-info/info-form.css'
 import FlyingStar from '@/components/background/FlyingStar';
