+++ on-fe/app/(auth)/login/page.js
@@ -1,5 +1,5 @@
 'use client';
-import LoginForm from '@/components/auth/LoginForm';
+import LoginForm from '@/components/form/LoginForm';
 import FormLayOut from '@/components/form/FormLayOut';
 import FlyingStar from '@/components/background/FlyingStar';
