+++ on-fe/app/room/create/page.jsx
@@ -4,12 +4,16 @@ import React from 'react'
 import Header from '@/components/Header'
 import FlyingStar from '@/components/background/FlyingStar'
 import DefaultPageFrame from '@/components/DefaultPageFrame'
+import FormLayOut from '@/components/form/FormLayOut'
 
 const page = () => {
   return (
    <DefaultPageFrame>
+      <FormLayOut>
+        
 
-    
+      </FormLayOut>
+
    </DefaultPageFrame>
   )
 }
